refactor(targets): split connector target persistence
